refactor: 배팅 금액 허용 범위 재설정
