+++ src/test/java/domain/MatchResultTest.java
@@ -1,0 +1,2 @@
+package domain;public class MatchResultTest {
+}
